requirements.txt
Django
Pillow
gunicorn
tensorflow-gpu==1.14.0
numpy==1.16.4
WhiteNoise
matplotlib
